eradicated most linter warnings and now everything compiles 100% warning-free in clang

+++ src/memory/boot.cpp
@@ -3,7 +3,6 @@
 #include <format>
 #include <fstream>
 #include <ios>
-#include <iostream>
 #include <stdexcept>
 #include <string>
 
@@ -19,7 +18,7 @@ BootROM::BootROM(const std::string &path) {
   file.seekg(0, std::ios::beg);
 
   rom_data.resize(rom_size);
-  if (!file.read(reinterpret_cast<char *>(rom_data.data()), rom_size))
+  if (!file.read(reinterpret_cast<char *>(rom_data.data()), static_cast<long long>(rom_size)))
     throw std::runtime_error("BootROM: Failed to fill buffer");
 }
